Base64 encode the service ID in the webhook path
We don't put any restrictions on the format of the Service ID so we need to
base64 it to prevent potentially losing data when it is inserted as a path
parameter (e.g. due to weird escaping rules).

Also, only remove webhooks if there was an old service to begin with.