Add stub RSS Service. Add DefaultService. Add Poller interface.
`DefaultService` serves as a useful no-op service to cut down on bloat when
implementing new Services. This means we can add more methods more freely to
the interface of `Service` without bogging down the Service implementation.

`Poller` is an interface which contains a polling interval and a function to
invoke. This will be used for RSS feeds.

Implemented a stub `RSSService`.