HACK: Have private/public fields for the webhook URL
We can't just use a public field because the caller may clobber it. If they
do clobber it, we've then lost the ability to set it back to what it was because
we don't store another copy anywhere. This patch fixes it by keeping a private
ref and always clobbering on Register().

This is horrible. We need to separate internal-storage and external-fields
better.