Add simple Mumble client to notify a room on join/leave
This adds a Mumble service that sits on a Mumble server and notifies a
Matrix room of users that join and leave. I'd like to expand this
functionality shortly (possibly including templates as found in the
Alertmanager service, or multiple Matrix rooms) but this has basic
functionality now.

I was sort of guessing at some of the API based on some of the
existing services, and I'm extremely new to Go. (This is the first Go
project I've contributed to!) Feedback on what I've inevitably done
horribly wrong is welcome. :)